show more events in homepage
there was a problem with the website where there would be more than 1 future events and it would show only one event on the main page. That might lead users to think that there is only one and not further explore. This commit adds a warning to the first page indicating that there are more events (in case there are of course)