correct index.html "upcoming event"
The upcoming event jumboton was showing the most distant event (in
the future). Now that has been fixed and simplified as it now shows
the event in the future closest to today (the next one).

Additionally, the feature that shows that there are more upcoming
events (previously implmented) now will make more sense as it indicates
events further into the future instead